requirements.txt
alpaca-py>=0.13.0
python-dotenv>=1.0.0
fastapi>=0.110.0
uvicorn>=0.29.0
jinja2>=3.1.0
pytz>=2024.1
httpx>=0.27.0
